+++ panel/resources/views/layouts/public.blade.php
@@ -29,7 +29,7 @@
                     <a href="{{ url('/#features') }}">Features</a>
                     <a href="{{ url('/docs') }}">Documentation</a>
                     <a href="{{ url('/docs/changelog') }}">Changelog</a>
-                    <a href="{{ url('/#about') }}">About</a>
+                    <a href="{{ url('/about') }}">About</a>
                     <a href="mailto:zilleali1245@gmail.com?subject=Zonclave%20inquiry" class="btn btn-primary btn-sm">Get in touch</a>
                 </nav>